backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1

openai-whisper==20231117
torch>=2.0.0
torchaudio>=2.0.0

sentence-transformers==2.6.1
scikit-learn==1.4.0
numpy==1.26.4

ffmpeg-python==0.2.0
python-multipart==0.0.9

pydantic==2.6.1
